+++ example/lib/pages/image_format_convert_example.dart
@@ -104,11 +104,6 @@ class _ImageFormatConvertExampleState extends State<ImageFormatConvertExample>
         },
         onCloseEditor: onCloseEditor,
       ),
-      configs: const ProImageEditorConfigs(
-        imageGenerationConfigs: ImageGeneratioConfigs(
-          allowEmptyEditCompletion: true,
-        ),
-      ),
     );
   }
 }