+++ lib/models/editor_callbacks/pro_image_editor_callbacks.dart
@@ -1,12 +1,13 @@
 // Flutter imports:
 import 'package:flutter/widgets.dart';
+import 'package:pro_image_editor/models/editor_callbacks/text_editor_callbacks.dart';
 
 // Project imports:
-import 'main_editor_callbacks.dart';
+import 'editor_callbacks_typedef.dart';
 
-export 'main_editor_callbacks.dart';
+export 'editor_callbacks_typedef.dart';
 
-/// A class representing configuration options for the Image Editor.
+/// A class representing callbacks for the Image Editor.
 class ProImageEditorCallbacks {
   /// A callback function that will be called when the editing is done,
   /// and it returns the edited image as a `Uint8List` with the format `png`.
@@ -55,6 +56,9 @@ class ProImageEditorCallbacks {
   /// The [ScaleEndDetails] parameter provides information about the scaling gesture.
   final Function(ScaleEndDetails)? onScaleEnd;
 
+  /// Callbacks from the text editor.
+  final TextEditorCallbacks? textEditorCallbacks;
+
   const ProImageEditorCallbacks({
     required this.onImageEditingComplete,
     this.onImageEditingStarted,
@@ -67,5 +71,6 @@ class ProImageEditorCallbacks {
     this.onScaleStart,
     this.onScaleUpdate,
     this.onScaleEnd,
+    this.textEditorCallbacks,
   });
 }